+++ RiskOfTheAncients2/Items/InfusedRaindrops.cs
@@ -1,7 +1,11 @@
 ﻿using BepInEx.Configuration;
+using Mono.Cecil.Cil;
+using MonoMod.Cil;
 using R2API;
 using RoR2;
 using System;
+using System.Collections.Generic;
+using System.ComponentModel;
 using UnityEngine;
 
 namespace ROTA2.Items
@@ -19,7 +23,67 @@ namespace ROTA2.Items
         public override void Hooks()
         {
             CharacterBody.onBodyInventoryChangedGlobal += OnInventoryChanged;
+            IL.RoR2.HealthComponent.TakeDamageProcess += (il) =>
+            {
+                ILCursor cursor = new(il);
+                cursor.GotoNext(
+                    x => x.MatchLdloc(0),
+                    x => x.MatchLdfld(out _),
+                    x => x.MatchLdfld(out _),
+                    x => x.MatchBrfalse(out _),
+                    x => x.MatchRet()
+                );
+                cursor.Emit(OpCodes.Ldarg_0);
+                cursor.Emit(OpCodes.Ldarg_1);
+                cursor.EmitDelegate<Action<HealthComponent, DamageInfo>>((self, info) =>
+                {
+                    int count = GetCount(self.body);
+                    if (info.rejected || info.damageType.damageType.HasFlag(DamageType.BypassBlock) || info.damage <= self.fullCombinedHealth * DamageThreshold / 100.0f)
+                    {
+                        if (count > 0)
+                        {
+                            counts[self.body].count = count;
+                        }
+
+                        return;
+                    }
+
+                    if (count > 0)
+                    {
+                        CountAndCharges value = counts[self.body];
+                        if (value.charges > 0)
+                        {
+                            info.damage -= DamageBlock;
+                            info.damage = Mathf.Max(0.0f, info.damage);
+                            if (info.damage == 0)
+                            {
+                                EffectData effectData = new EffectData
+                                {
+                                    origin = info.position,
+                                    rotation = Util.QuaternionSafeLookRotation((info.force != Vector3.zero) ? info.force : UnityEngine.Random.onUnitSphere)
+                                };
+                                EffectManager.SpawnEffect(HealthComponent.AssetReferences.bearEffectPrefab, effectData, transmit: true);
+                                info.rejected = true;
+                            }
+
+                            --value.charges;
+                            if (value.charges % BlockCount == 0)
+                            {
+                                ignore = true;
+                                self.body.inventory.RemoveItem(ItemDef);
+                                self.body.inventory.GiveItem(DehydratedRaindrops.Instance.ItemDef);
+                                ignore = false;
+                                --value.count;
+
+                                CharacterMasterNotificationQueue.PushItemTransformNotification(self.body.master, ItemDef.itemIndex, DehydratedRaindrops.Instance.ItemDef.itemIndex, CharacterMasterNotificationQueue.TransformationType.Default);
+                                Util.PlaySound("InfusedRaindrops", self.body.gameObject);
+                            }
+                        }
+                    }
+                });
+            };
         }
+
         public override void Init(ConfigFile configuration)
         {
             CreateConfig(configuration);
@@ -38,94 +102,85 @@ namespace ROTA2.Items
             BlockCount      = configuration.Bind("Item: " + ItemName, "Block Count Per Stack", 10, "").Value;
         }
 
+        public class CountAndCharges
+        {
+            public int count;
+            public int charges;
+        }
+        public Dictionary<CharacterBody, CountAndCharges> counts = [];
+        private bool ignore = false;
         private void OnInventoryChanged(CharacterBody body)
         {
-            int count = GetCount(body);
-            if (GetCount(body) > 0 && !body.GetComponent<InfusedRaindropsBehavior>())
+            if (ignore)
             {
-                body.gameObject.AddComponent<InfusedRaindropsBehavior>();
+                return;
+            }
+
+            int count = GetCount(body);
+            if (GetCount(body) > 0)
+            {
+                if (!counts.ContainsKey(body))
+                {
+                    counts.Add(body, new ()
+                    {
+                        charges = BlockCount * count,
+                        count = count
+                    });
+
+                    return;
+                }
+                if (counts[body].count < count)
+                {
+                    counts[body].charges += BlockCount * (count - counts[body].count);
+                    counts[body].count = count;
+                }
+                else if (counts[body].count > count)
+                {
+                    counts[body].charges = BlockCount * count;
+                    counts[body].count = count;
+                }
             }
         }
-
-        public class InfusedRaindropsBehavior : MonoBehaviour
+        private void OnTakeDamage(On.RoR2.HealthComponent.orig_TakeDamage orig, HealthComponent self, DamageInfo info)
         {
-            CharacterBody body;
-            public int remaining_charges;
-            int last_count;
-            private bool ignore;
-
-            public void Awake()
+            int count = GetCount(self.body);
+            if (!info.rejected && info.damage > self.fullCombinedHealth * DamageThreshold / 100.0f && count > 0)
             {
-                body = GetComponent<CharacterBody>();
-                last_count = InfusedRaindrops.Instance.GetCount(body);
-                remaining_charges = last_count * InfusedRaindrops.Instance.BlockCount;
-                ignore = false;
-
-                On.RoR2.CharacterBody.OnInventoryChanged += OnInventoryChanged;
-                On.RoR2.HealthComponent.TakeDamage += OnTakeDamage;
-            }
-
-            private void OnInventoryChanged(On.RoR2.CharacterBody.orig_OnInventoryChanged orig, CharacterBody body)
-            {
-                if (ignore || body != this.body)
+                if (counts[self.body].charges > 0)
                 {
-                    orig(body);
-                    return;
-                }
-
-                int count = InfusedRaindrops.Instance.GetCount(body);
-                if (count > last_count)
-                {
-                    remaining_charges += (count - last_count) * InfusedRaindrops.Instance.BlockCount;
-                    last_count = count;
-                }
-                else if (count < last_count)
-                {
-                    remaining_charges -= count * InfusedRaindrops.Instance.BlockCount;
-                    last_count = count;
-                }
-
-                orig(body);
-            }
-            private void OnTakeDamage(On.RoR2.HealthComponent.orig_TakeDamage orig, HealthComponent self, DamageInfo info)
-            {
-                if (info.rejected || !self || self.body != body || remaining_charges <= 0)
-                {
-                    orig(self, info);
-                    return;
-                }
-
-                if (info.damage > self.fullCombinedHealth * InfusedRaindrops.Instance.DamageThreshold / 100.0f)
-                {
-                    info.damage = Math.Max(0.0f, info.damage - InfusedRaindrops.Instance.DamageBlock);
-                    if (info.damage == 0.0f)
+                    info.damage -= DamageBlock;
+                    info.damage = Mathf.Max(0.0f, info.damage);
+                    if (info.damage == 0)
                     {
-                        EffectData effect = new EffectData
+                        EffectData effectData = new EffectData
                         {
                             origin = info.position,
                             rotation = Util.QuaternionSafeLookRotation((info.force != Vector3.zero) ? info.force : UnityEngine.Random.onUnitSphere)
                         };
-                        EffectManager.SpawnEffect(HealthComponent.AssetReferences.bearEffectPrefab, effect, transmit: true);
+                        EffectManager.SpawnEffect(HealthComponent.AssetReferences.bearEffectPrefab, effectData, transmit: true);
                         info.rejected = true;
                     }
 
-                    --remaining_charges;
-                    if (remaining_charges % 10 == 0)
+                    --counts[self.body].charges;
+                    if (counts[self.body].charges % 10 == 0)
                     {
                         ignore = true;
-                        body.inventory.RemoveItem(InfusedRaindrops.Instance.ItemDef, 1);
-                        body.inventory.GiveItem(DehydratedRaindrops.Instance.ItemDef, 1);
+                        self.body.inventory.RemoveItem(ItemDef);
+                        self.body.inventory.GiveItem(DehydratedRaindrops.Instance.ItemDef);
                         ignore = false;
+                        --counts[self.body].count;
 
-                        CharacterMasterNotificationQueue.PushItemTransformNotification(body.master, InfusedRaindrops.Instance.ItemDef.itemIndex, DehydratedRaindrops.Instance.ItemDef.itemIndex, CharacterMasterNotificationQueue.TransformationType.Default);
-                        body.RecalculateStats();
-
-                        Util.PlaySound("InfusedRaindrops", body.gameObject);
+                        CharacterMasterNotificationQueue.PushItemTransformNotification(self.body.master, ItemDef.itemIndex, DehydratedRaindrops.Instance.ItemDef.itemIndex, CharacterMasterNotificationQueue.TransformationType.Default);
+                        Util.PlaySound("InfusedRaindrops", self.body.gameObject);
                     }
                 }
-
-                orig(self, info);
             }
+            else if (count > 0)
+            {
+                counts[self.body].count = count;
+            }
+
+            orig(self, info);
         }
     }